+++ CHANGELOG.md
@@ -65,6 +65,8 @@ CHANGELOG
 - Add prompt for availability zone in pcluster configure automated subnets creation.
 - Add configuration HeadNode.Imds.Secured to enable/disable restricted access to IMDS.
 - Use different permissions in instance roles based on the scheduler and the node's role in the cluster.
+- Support restart/reboot for instance type with instance store (ephemeral drives).
+- Remove instance store software encryption option (encrypted_ephemeral).
 - Implement scaling protection mechanism with Slurm scheduler: compute fleet is automatically set to 'PROTECTED' state
   in case recurrent failures are encountered when provisioning nodes.
 - Upgrade Slurm to version 20.11.8.